Use local etcd for dynamic configuration.
Enable the use of etcdctl to set and change /live/downstreams
dynamically, which will reload nginx inside the RTMP container.

Bind etcd to the docker0 and lo0 addresses instead of any public address
for security reasons.

This currently requires knowledge of the special Docker 1.9+ IP address
for docker0, which may change and should be replaced with a mechanism to
look up the IP.

Remove cors_http_origin variable since it is no longer required.